% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.cov}
\alias{create.cov}
\title{Calculate the covariance matrix}
\usage{
create.cov(x, knots)
}
\arguments{
\item{Vector}{\code{x} the explanatory variables}

\item{Vector}{\code{knots} the knots to be used}
}
\value{
Matrix containing the covariates for linear spline regression
}
\description{
This is a helper function to calculated the covariance
matrix for the linear regression spline function. This
function is not to be exported.
}
\author{
Creagh Briercliffe
}
\keyword{array}

