% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{linearspline}
\alias{linearspline}
\alias{linearspline-package}
\title{R package: linearspline}
\description{
Fit linear regression splines to numeric data.
}
\details{
Linear spline regression works by fitting piece-wise linear
functions to segments of the data between connecting points,
called knots. One benefit of linear spline regression is that
it can produce a better local fit of the data between the knots.
}
\author{
Creagh Briercliffe
}

