% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lmspline}
\alias{lmspline}
\title{Linear spline regression}
\usage{
lmspline(x, y, nknots = 1, na.rm = FALSE)
}
\arguments{
\item{Vector}{\code{x} of the numeric valued data}

\item{Vector}{\code{y} of the responses}

\item{Numeric}{\code{nknots} a positive integer corresponding
to the number of knots}

\item{Logical}{\code{na.rm} takes value \code{TRUE} if \code{NA}
values should be ignored}
}
\value{
List containing the fitted model and the vector of knots used.
}
\description{
This function returns a list of coefficients and knots used
to fit a linear spline to the given data. Given a number of knots,
 \code{nknots}, this function will place \code{nknots} knots at
 evenly spaced quantiles of the data, and fit a linear regression spline.
}
\details{
Linear spline regression works by fitting piece-wise linear
functions to segments of the data between connecting points,
called knots. One benefit of linear spline regression is that
it can produce a better local fit of the data between the knots.
}
\examples{
# Example 1
x <- rnorm(100, 0, 1)
y <- 3 + pmin(2* x, 0) + rnorm(1, 0, 1)
lmspline(x, y)

# Example 2
lmspline(x, y, nknots = 3, na.rm = FALSE)

# Example 3
gd_url <- "http://tiny.cc/gapminder"
gDat <- read.delim(file = gd_url)
lmspline(gDat$gdpPercap, gDat$lifeExp, 2, na.rm = TRUE)
}
\author{
Creagh Briercliffe
}
\keyword{models}

