% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pred.spline}
\alias{pred.spline}
\title{Predict linear spline}
\usage{
pred.spline(x, lspline)
}
\arguments{
\item{Vector}{\code{x} of explanatory values to predict from}

\item{Object}{\code{lspline} an lmspline object that contains
the fitted linear spline}
}
\value{
Vector \code{y} the predictions for the given data and
linear spline fit
}
\description{
This function accepts a linear spline model and a set of explanatory
variables, for which it returns the predictions.
}
\examples{
# Example 1

x <- rnorm(100, 0, 1)
y <- 3 + pmin(2* x, 0) + rnorm(1, 0, 1)
lspline <- lmspline(x, y, nknots = 1)

pred.spline(x, lspline)

# Example 2 - plotting with pred.spline

library(ggplot2)
df <- data.frame(x,y)

ggplot(df, aes(x, y)) + geom_point() +
stat_function(fun = pred.spline, args = list(lspline), colour = 'blue', size = 0.7)
}
\author{
Creagh Briercliffe
}
\keyword{datagen}

