% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lmspline}
\alias{lmspline}
\title{Linear spline regression}
\usage{
lmspline(x, y, num_knots = 1, na.rm = FALSE)
}
\arguments{
\item{Vector}{\code{x} of the numeric valued data}

\item{Vector}{\code{y} of the responses}

\item{Numeric}{\code{num_knots} a positive integer corresponding
to the number of knots}

\item{Logical}{\code{na.rm} takes value \code{TRUE} if \code{NA}
values should be ignored}
}
\value{
List containing the fitted model and the vector of knots used.
}
\description{
This function returns a list of coefficients and knots used
to fit a linear spline to the given data.
}
\details{
Linear spline regression works by fitting piece-wise linear
functions to segments of the data between connecting points,
called knots. One benefit of linear spline regression is that
it can produce a better local fit of the data between the knots.
}
\examples{
lmspline(x, y)
lmspline(num_knots = 3, x, y, na.rm = FALSE)
}
\author{
Creagh Briercliffe
}
\keyword{models}

